backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
langchain
langchain-community
langchain-ollama
ChatOllama
python-dotenv
faiss-cpu
tiktoken 
